bugfix: S3C-4425 CRR failed: don't wait for delivery report
Don't wait for the Kafka delivery reports to be received for the
backbeat-replication-failed queue, instead, move on immediately with
the processing of further items.

This solves a low throughput issue when there are a significant number
of FAILED entries because each of them held a queue worker for a few
seconds, preventing new status updates from being done.
